src/main/webapp/app/entities/affiliate/route/affiliate-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IAffiliate, Affiliate } from '../affiliate.model';
import { AffiliateService } from '../service/affiliate.service';

import { AffiliateRoutingResolveService } from './affiliate-routing-resolve.service';

describe('Affiliate routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: AffiliateRoutingResolveService;
  let service: AffiliateService;
  let resultAffiliate: IAffiliate | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(AffiliateRoutingResolveService);
    service = TestBed.inject(AffiliateService);
    resultAffiliate = undefined;
  });

  describe('resolve', () => {
    it('should return IAffiliate returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAffiliate = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAffiliate).toEqual({ id: 123 });
    });

    it('should return new IAffiliate if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAffiliate = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultAffiliate).toEqual(new Affiliate());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Affiliate })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAffiliate = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAffiliate).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
